requirements.txt
asn1crypto==0.24.0
backcall==0.1.0
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
colorama==0.3.9
copyexcel==0.1
copyexcel1==0.1
cryptography==2.2.2
decorator==4.3.0
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.0.6
django-allauth==0.36.0
django-cors-headers==2.3.0
django-debug-toolbar==1.9.1
django-filter==1.1.0
django-rest-auth==0.9.3
django-rest-framework==0.1.0
django-webpack-loader==0.6.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==3.2.3
dumptoexcel==0.1
et-xmlfile==1.0.1
facebook-sdk==2.0.0
gunicorn==19.9.0
idna==2.7
ipdb==0.11
ipython==6.4.0
ipython-genutils==0.2.0
jdcal==1.4
jedi==0.12.0
mail-task==0.1
mailer==0.8.1
Markdown==2.6.11
mysql-connector-python==8.0.11
mysql-mini==1.0
mysqlclient==1.3.12
oauthlib==2.1.0
openpyxl==2.5.3
packaging==17.1
parso==0.2.1
permission==0.4.1
pickleshare==0.7.4
Pillow==5.1.0
Pillow-PIL==0.1.dev0
pip-review==1.0
prompt-toolkit==1.0.15
protobuf==3.6.0
psutil==5.4.6
psycopg2==2.7.5
pycparser==2.18
Pygments==2.2.0
PyJWT==1.5.3
PyMySQL==0.9.0
pyparsing==2.2.0
python3-openid==3.1.0
pytz==2017.2
requests==2.19.1
requests-oauthlib==1.0.0
simplegeneric==0.8.1
six==1.11.0
sqlparse==0.2.4
strcmd==0.1
traitlets==4.3.2
urllib3==1.23
virtualenv==15.2.0
wcwidth==0.1.7
whitenoise==3.3.1
work6==0.1
